moves stateful variables into their own folder
